requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
apscheduler==3.10.4
python-dotenv==1.0.0
playwright==1.49.0
nest-asyncio==1.6.0
